cad-pulse: better select the output port when selecting voice call mode
Forcing the earpiece port can be a wrong idea when e.g. the person is
wearing headphones: we want to keep outputting sound through the
headphones instead of suddenly switching to the earpiece.

This is done by retrieving the highest priority available port which
isn't the speaker. It requires priorities to be set right and having no
other output than speaker, earpiece or headphones, but for now it's a
resonable constraint.