WIP: pulse: support devices running pulseaudio-modules-droid
The underlying Android HAL switches modes once the next routing change happens.
Thus, we need to "park" the sink/source before switching the actual port,
which is not needed on native devices.

This commit adds detection for sinks and sources created by pulseaudio-modules-droid,
and handles port parking if needed.

This also means that, on droid devices, port changes happen also when a voicecall profile
has been detected.

Note: this is work in progress and as such this commit might (will) be reworked, split in
multiple parts so please do not merge.

Also error handling for operations beyond the last one is currently missing :)

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>