cad-pulse: filter card based on available ports
The card name and main attributes aren't always enough to properly
identify the proper card to use. For example, the HDMI audio output on
PP is named `platform-hdmi-sound`, and is sometimes picked up as the
main card even though it's obviously wrong.

With this commit, only the cards having both a `Speaker` and a
`Earpiece` ports will be considered, which should solve most of the
remaining problems.

We also setup a retry timer in case no suitable card is found in an
attempt to fix #18.