cad-pulse: track current state
As all actual operations are performed by the backend, it is the best
place to track current state. This commit ensures `CadPulse` can
properly determine it on startup and updates it whenever an operation
is executed.

This requires a small additional change to `cad-manager.c` so the
operation data aren't freed in the callback, but only after they have
been fully processed by the `CadPulse` object.