cad-pulse: simplify mic mute operation
Now that CadPulse tracks the routing state, there's no need to retrieve
the source info prior to (un)muting the mic, we can simply call the
pulseaudio mute function directly. It has the added benefit of
preventing a race condition when switching to default profile while
muted.

Similar changes would be useful for other operations, but they require
refactoring the ports handling first, so let's stick with the easy one
for now.