cad-pulse: Don't crash on NULL cad_op callback
When CallAudioOperation fails it invokes the set callback which usually
points to `complete_command_cb()` to return the result to DBus. However
cad_pulse_select_mode() invokes cad_pulse_mute_mic() without a callback
set hence we crash in the error path like:

 #1  0x0000aaaad7949aac in cad_pulse_mute_mic (mute=0, cad_op=0xaaab15802f40) at ../src/cad-pulse.c:1097
 #2  0x0000aaaad7949c5c in cad_pulse_select_mode (mode=CALL_AUDIO_MODE_DEFAULT, cad_op=0xaaab15809730) at ../src/cad-pulse.c:950
 #3  0x0000aaaad7947b3c in cad_manager_handle_select_mode (object=0xaaab157e98d0, invocation=0xffffa4009c40, mode=0) at ../src/cad-manager.c:90
 ...

Guard against that case. Do so in other call sites as well so
we don't hit the same problem should we invoke those functions
internally as well.

To reproduce run `callaudiocli -m 0` on a system that does't
detect a suitable card.