cad-pulse: also accept "Handset" type input ports
The current UCM config for the Librem 5 has the bottom microphone named
`Handset2`, causing PulseAudio to set the "Handset" type for it. As a
consequence, it is never selected by `callaudiod` which only selects the
stereo microphone, causing echo on the other side of the call due to the
top micrphone being located next to the earpiece speaker.

This change will make the bottom microphone the preferred input port
for voice calls, making it more pleasant to users.