cad-pulse: don't filter input ports based on type
This can be too limiting in the long run as we should also account for
USB devices, S/PDIF inputs maybe, and so on... This change was
introduced in 89d95738 as an additional security but it's probably
over-cautious. If we were to re-introduce such a filtering, it should be
implemented the other way around: exclude unwanted port types instead
of having to whitelist every single compatible input type.